+++ src/main/scala/io/viash/functionality/arguments/FileArgument.scala
@@ -74,14 +74,25 @@ case class FileArgument(
       "yaml")
   default: OneOrMore[Path] = Nil,
 
-  @description("The file or folder should exist before the start of execution. If set to `true`, an error will be produced if the file or folder wasn't found.")
+  @description("Checks whether the file or folder exists. For input files, this check will happen " +
+    "before the execution of the script, while for output files the check will happen afterwards.")
   @example(
     """- name: --my_file
       |  type: file
       |  must_exist: true
       |""".stripMargin,
       "yaml")
-  must_exist: Boolean = false,
+  must_exist: Boolean = true,
+
+  @description("If the output filename is a path and it does not exist, create it before executing the script (only for `direction: output`).")
+  @example(
+    """- name: --my_file
+      |  type: file
+      |  direction: output
+      |  create_parent: true
+      |""".stripMargin,
+      "yaml")
+  create_parent: Boolean = true,
 
   @description("Make the value for this argument required. If set to `true`, an error will be produced if no value was provided. `false` by default.")
   @example(
@@ -139,6 +150,10 @@ case class FileArgument(
     multiple_sep: String,
     dest: String
   ): Argument[Path] = {
-    copy(name, alternatives, description, info, example, default, this.must_exist, required, direction, multiple, multiple_sep, dest, `type`)
+    copy(
+      name, alternatives, description, info, example, default, 
+      this.must_exist, this.create_parent, required, direction, 
+      multiple, multiple_sep, dest, `type`
+    )
   }
 }