+++ src/test/resources/testscala/test.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 set -ex
 
+echo foo > NOTICE
+echo foo > resource2.txt
+
 echo ">>> Checking whether output is correct"
 ./testscala "NOTICE" --real_number 10.5 --whole_number=10 -s "a string with spaces" --truth \
   --optional foo --optional_with_default bar \
@@ -43,7 +46,7 @@ $meta_executable \
   --whole_number=789 \
   -s "a \\ b \$ c \` d \" e ' f \n g # h @ i { j } k \"\"\" l ''' m todo_add_back_DOLLAR_VIASH_TEMP n : o ; p" \
   ---cpus 666 \
-  ---memory 1GB \
+  ---memory 100PB \
   > output2.txt
 
 [[ ! -f output2.txt ]] && echo "Output file could not be found!" && exit 1