+++ src/test/scala/io/viash/e2e/ns_build/MainNSBuildNativeSuite.scala
@@ -1,4 +1,6 @@
-package io.viash
+package io.viash.e2e.ns_build
+
+import io.viash._
 
 import io.viash.config.Config
 import io.viash.helpers.{Exec, IO}
@@ -30,7 +32,7 @@ class MainNSBuildNativeSuite extends FunSuite with BeforeAndAfterAll{
 
   // convert testbash
   test("viash ns can build") {
-    val (stdout, stderr) = TestHelper.testMainWithStdErr(
+    val (stdout, stderr, exitCode) = TestHelper.testMainWithStdErr(
     "ns", "build",
       "-s", nsPath,
       "-t", tempFolStr
@@ -38,6 +40,7 @@ class MainNSBuildNativeSuite extends FunSuite with BeforeAndAfterAll{
 
     assert(nsFolder.exists)
     assert(nsFolder.isDirectory)
+    assert(exitCode == 0)
 
     for ((component, _, _, _) <- components) {
       val executable = componentExecutableFile(component)