+++ src/test/scala/io/viash/e2e/ns_exec/MainNSExecNativeSuite.scala
@@ -1,4 +1,6 @@
-package io.viash
+package io.viash.e2e.ns_exec
+
+import io.viash._
 
 import io.viash.helpers.IO
 import org.scalatest.{BeforeAndAfterAll, FunSuite}
@@ -23,7 +25,7 @@ class MainNSExecNativeSuite extends FunSuite with BeforeAndAfterAll {
   private val tempFolStr = temporaryFolder.toString
 
   test("Check whether ns exec \\; works") {
-    val (stdoutRaw, stderrRaw) =
+    val (stdoutRaw, stderrRaw, _) =
       TestHelper.testMainWithStdErr(
         "ns", "exec",
         "--src", nsPath,
@@ -42,7 +44,7 @@ class MainNSExecNativeSuite extends FunSuite with BeforeAndAfterAll {
   }
 
   test("Check whether ns exec + works") {
-    val (stdoutRaw, stderrRaw) = TestHelper.testMainWithStdErr(
+    val (stdoutRaw, stderrRaw, _) = TestHelper.testMainWithStdErr(
       "ns", "exec",
       "--src", nsPath,
       "echo {path} +"